refactoring dan riwayat keuangan transaksi

+++ database/migrations/2019_02_13_045535_create_table_transaksi_pelunasan.php
@@ -15,7 +15,7 @@ class CreateTableTransaksiPelunasan extends Migration
     {
         Schema::create('transaksi__pelunasan', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('transaksi_id');
+            $table->unsignedBigInteger('transaksi_id');
             $table->timestamp('tanggal')->nullable();
             $table->bigInteger('nilai');
             $table->bigInteger('debit')->default(0);
